test: dedupe manpage checks
